update:/ code destructure with iUser types file
