+++ services/user/type.d.ts
@@ -1,3 +1,19 @@
+import {Document} from "mongoose";
+
+export interface iUser extends Document {
+    name: string;
+    email: string;
+    password: string;
+    avatar: {
+        public_id: string;
+        ur: string;
+    };
+    role: string;
+    isVerified: boolean;
+    courses: Array<{courseId: string}>;
+    comparePassword: (password: string) => Promise<boolean>;
+}
+
 export interface iRegistrationBody {
     name: string;
     email: string;
